Switched to using optimizing a multi-objective
I want to minimize both the failure_rate and the repetition_penalty. Ideally, we wouldn't need to se the repetition_penalty since it might have unintended consequences. Plus earlier runs showed various configurations that had a failure rate of 0.0. This way we can say which of those is better.

Also started initial blog. It will need to be rewritten