Update with final results + clean up code
Phase 2 results finished. New plots and updated README

Changed codes to enable the newer two phase approach in a single script.  The phase2 version was mostly just a copy of the first. I didn't use this generate the results shown in the images and README, but will use it for investigating fp8 version of Llama-3.1